requirements.txt
torch==2.4.0
torchvision==0.19.0
accelerate==0.31.0
diffusers==0.27.2
matplotlib==3.9.1
numpy==1.26.4
opencv_python==4.10.0.84
pillow==10.3.0
PyYAML==6.0.1
scipy==1.13.1
setuptools==51.0.0
scikit-image==0.24.0
tqdm==4.66.4
transformers==4.46.3
fvcore==0.1.5.post20221221
cloudpickle==3.0.0
omegaconf==2.3.0
pycocotools==2.0.8
av==12.3.0
gradio==4.41.0
peft==0.14.0
huggingface_hub==0.25.0
uvicorn
fastapi
pydantic